% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_trajs}
\alias{simulate_trajs}
\title{Simulate infectious disease trajectories}
\usage{
simulate_trajs(n_simu, ntot, n_days, n0, n_days_inf, prob_infect)
}
\arguments{
\item{n_simu}{integer, number of simulated trajectories}

\item{ntot}{integer,  total number of individuals}

\item{n_days}{integer, number of days in trajectory}

\item{n0}{integer,  number infected on day 1}

\item{n_days_inf}{integer, for someone infected on day t,
we assume that they are infectious on day (t, t+n_days_inf)}

\item{prob_infect}{numeric in (0,1), probability that a susceptible individual
gets infected is n_infectious/ntot*prob_infect}
}
\value{
list of simulated trajectories
}
\description{
Simulate infectious disease trajectories
}
